# farmasi-service/app/Http/Resources/PharmacyResource.php
<?php
namespace App\Http\Resources;
use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;
class PharmacyResource extends JsonResource
{
    public function toArray(Request $request): array
    {
        return [
            'id'             => $this->id,
            'patient_id'     => $this->patient_id,
            'appointment_id' => $this->appointment_id,
            'medicine_name'  => $this->medicine_name,
            'dosage'         => $this->dosage,
            'frequency'      => $this->frequency,
            'quantity'       => $this->quantity,
            'instructions'   => $this->instructions,
            'status'         => $this->status,
            'created_at'     => $this->created_at,
            'updated_at'     => $this->updated_at,
        ];
    }
}